requirements.txt
streamlit
langchain
langchain_community
transformers
PyPDF2
pdfplumber
dotenv
docx
ipykernel
google-generativeai
protobuf==3.20.3